Get basic tests working again and as standalone as possible, though keep around the old ambitious stuff in emulator-collage. Switch to concept of 'machine checksum' including cpu and ram. Some modernization and cleanup.
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@4548 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ fceulib/test-util.cc
@@ -54,10 +54,10 @@ vector<string> SplitToLines(const string &s) {
 }
 
 string Chop(string &line) {
-  for (int i = 0; i < line.length(); i ++) {
+  for (int i = 0; i < (int)line.length(); i ++) {
     if (line[i] != ' ') {
       string acc;
-      for(int j = i; j < line.length(); j ++) {
+      for (int j = i; j < (int)line.length(); j ++) {
         if (line[j] == ' ') {
           line = line.substr(j, line.length() - j);
           return acc;
@@ -74,7 +74,7 @@ string Chop(string &line) {
 
 string LoseWhiteL(const string &s) {
   for (unsigned int i = 0; i < s.length(); i ++) {
-    switch(s[i]) {
+    switch (s[i]) {
     case ' ':
     case '\n':
     case '\r':
